// SGUEES-SPA/src/app/pages/General/gen-rubro/models/gen-rubro.ts
export interface GenRubro {
	CORR_EMPRESA: number;
	CORR_RUBRO: number;
	NOMBRE_RUBRO: string;
	DESCRIPCION_RUBRO: string;
	ES_IMPUESTO: boolean;
	POR_IMPUESTO: number;
	MUESTRA_DETALLE: boolean;
	MUESTRA_TOTAL: boolean;
	SUMA_RESTA: number;
	NOMBRE_SUMA_RESTA: string;
	CLASE_RUBRO: string;
	NOMBRE_CLASE_RUBRO: string;
	TIPO_APLICACION: string;
	NOMBRE_TIPO_APLICACION: string;
}